move enableAllSubCategories to DisplayStyleState
